Move loss functions to their own file
