+++ main.py
@@ -7,6 +7,7 @@ from torchaudio.transforms import MelSpectrogram
 from tqdm import tqdm
 
 from net import SoundStream, WaveDiscriminator, STFTDiscriminator
+from loss import adversarial_g_loss, feature_loss, adversarial_d_loss
 from dataset import NSynthDataset
 
 
@@ -27,6 +28,7 @@ soundstream.to(device)
 wave_disc.to(device)
 stft_disc.to(device)
 
+
 def collate_fn(batch):
     lengths = torch.tensor([elem.shape[-1] for elem in batch])
     return nn.utils.rnn.pad_sequence(batch, batch_first=True), lengths
@@ -41,23 +43,6 @@ valid_loader = DataLoader(valid_dataset, batch_size=BATCH_SIZE, collate_fn=colla
 test_dataset = NSynthDataset(audio_dir="./data/nsynth-test.jsonwav/nsynth-test/audio")
 test_loader = DataLoader(test_dataset, batch_size=BATCH_SIZE, collate_fn=collate_fn, num_workers=2)
 
-def adversarial_g_loss(features_stft_disc_G_x, features_wave_disc_G_x, lengths_stft, lengths_wave):
-    wave_disc_names = lengths_wave.keys()
-    
-    stft_loss = F.relu(1-features_stft_disc_G_x[-1]).sum(dim=3).squeeze()/lengths_stft[-1].squeeze()
-    wave_loss = torch.cat([F.relu(1-features_wave_disc_G_x[key][-1]).sum(dim=2).squeeze()/lengths_wave[key][-1].squeeze() for key in wave_disc_names])
-    loss = torch.cat([stft_loss, wave_loss]).mean()
-    
-    return loss
-
-def feature_loss(features_stft_disc_x, features_wave_disc_x, features_stft_disc_G_x, features_wave_disc_G_x, lengths_wave, lengths_stft):
-    wave_disc_names = lengths_wave.keys()
-    
-    stft_loss = torch.stack([((feat_x-feat_G_x).abs().sum(dim=-1)/lengths_stft[i].view(-1,1,1)).sum(dim=-1).sum(dim=-1) for i, (feat_x, feat_G_x) in enumerate(zip(features_stft_disc_x, features_stft_disc_G_x))], dim=1).mean(dim=1, keepdim=True)
-    wave_loss = torch.stack([torch.stack([(feat_x-feat_G_x).abs().sum(dim=-1).sum(dim=-1)/lengths_wave[key][i] for i, (feat_x, feat_G_x) in enumerate(zip(features_wave_disc_x[key], features_wave_disc_G_x[key]))], dim=1) for key in wave_disc_names], dim=2).mean(dim=1)
-    loss = torch.cat([stft_loss, wave_loss], dim=1).mean()
-    
-    return loss
 
 def spectral_reconstruction_loss(x, G_x, eps=1e-4):
     L = 0
@@ -72,19 +57,6 @@ def spectral_reconstruction_loss(x, G_x, eps=1e-4):
         L += loss
     
     return L
-
-def adversarial_d_loss(features_stft_disc_x, features_wave_disc_x, features_stft_disc_G_x, features_wave_disc_G_x, lengths_stft, lengths_wave):
-    wave_disc_names = lengths_wave.keys()
-    
-    real_stft_loss = F.relu(1-features_stft_disc_x[-1]).sum(dim=3).squeeze()/lengths_stft[-1].squeeze()
-    real_wave_loss = torch.stack([F.relu(1-features_wave_disc_x[key][-1]).sum(dim=-1).squeeze()/lengths_wave[key][-1].squeeze() for key in wave_disc_names], dim=1)
-    real_loss = torch.cat([real_stft_loss.view(-1,1), real_wave_loss], dim=1).mean()
-    
-    generated_stft_loss = F.relu(1+features_stft_disc_G_x[-1]).sum(dim=-1).squeeze()/lengths_stft[-1].squeeze()
-    generated_wave_loss = torch.stack([F.relu(1+features_wave_disc_G_x[key][-1]).sum(dim=-1).squeeze()/lengths_wave[key][-1].squeeze() for key in wave_disc_names], dim=1)
-    generated_loss = torch.cat([generated_stft_loss.view(-1,1), generated_wave_loss], dim=1).mean()
-    
-    return real_loss + generated_loss
 
 optimizer_g = optim.Adam(soundstream.parameters(), lr=1e-4, betas=(0.5, 0.9))
 optimizer_d = optim.Adam(list(wave_disc.parameters()) + list(stft_disc.parameters()), lr=1e-4, betas=(0.5, 0.9))
